requirements.txt
cloudpickle
pandas
numpy
matplotlib
seaborn
PyYAML
ipykernel